fix($Fix): Fix signature with ECDSA, add btcec deps in glide.yaml, remove cgo libsecp256k1.
